[spr] initial version
Created using spr 1.3.6-beta.1
